requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-docx==1.1.0
PyPDF2==3.0.1
pypandoc==1.12
openai==1.6.1
anthropic==0.18.1
python-dotenv==1.0.0
werkzeug==3.0.1
gunicorn==21.2.0
markdown==3.5.1
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.4
numpy==1.26.2
pdfplumber==0.10.3
Pillow==10.1.0
mammoth==1.6.0
requests==2.31.0
reportlab==4.0.7
